from_gcs_to_bq_service_industry/requirements.txt
google-cloud-storage
google-cloud-bigquey